Send the default thinking budget as an integer
Found in the pre-release audit. Reading the documented config key (f63cde0)
made ANTHROPIC_DEFAULT_THINKING_BUDGET reachable for the first time, and env()
returns a string, so budget_tokens went out as "3000" rather than 3000. It is
cast to an integer, and the test covers the string an environment variable
gives.

+++ src/Providers/Anthropic/Concerns/ResolvesThinking.php
@@ -52,9 +52,11 @@ trait ResolvesThinking
         if (($thinking['enabled'] ?? null) === true) {
             return [
                 'type' => 'enabled',
+                // Cast: ANTHROPIC_DEFAULT_THINKING_BUDGET arrives from the
+                // environment as a string, and budget_tokens is an integer.
                 'budget_tokens' => is_int($thinking['budgetTokens'] ?? null)
                     ? $thinking['budgetTokens']
-                    : config('prism.providers.anthropic.default_thinking_budget', 1024),
+                    : (int) config('prism.providers.anthropic.default_thinking_budget', 1024),
             ];
         }
